requirements.txt
playwright
pandas
pydantic-settings
pyyaml
ws-api
pdfplumber
monopoly-core
python-dotenv
